Function a_fractal_a_day::postprocess_image_for_twitter

pub fn postprocess_image_for_twitter(input: &str, outfile: &str)

reduce the size to something twitter allows

Arguments

input - path to the source png

output - path where the postprocessed png should be saved

Remarks

We need to ensure that the filesize is <= 5 MiB. therefore, we will shrink by 10%, measure the size and repeat, until we reach the limit.

Note: If imagemagick’s convert is not in the path, this function will do nothing, but logging an error.