Module a_fractal_a_day::histogram

Structs

ColoredHistogram
data structure containing a 2d-histogram with 4 channels (rgba)

Enums

BoundsTypes

Functions

bounds
find bounds containing all input points

bounds_without_outliers
finds bounds disarding a number of outliers in each direction

bounds_zoom
finds bounds centering on the median filling the given aspect ratio

histogram
generates a 2d-histogram from an iterator

histogram1d